updated cacheprovider to support validation and multi cache provider

+++ src/CacheProvider.Multi/MultiCacheProvider.cs
@@ -1,4 +1,5 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
 using System.Collections.Specialized;
 using System.Configuration;
 using System.Globalization;
@@ -14,12 +15,14 @@ namespace CacheProvider.Multi
     {
         private bool _isEnabled;
         public readonly List<ICacheProvider> CacheProviders;
+        private ICacheProvider _masterCacheProvider;
 
         public MultiCacheProvider()
         {
             CacheProviders = new List<ICacheProvider>();
         }
 
+        #region init
         /// <summary>
         ///     Initialize from config
         /// </summary>
@@ -69,7 +72,7 @@ namespace CacheProvider.Multi
                     timeSlice = 1;
                 }
                 valueCollection["timeout"] = (cacheExpirationTime / timeSlice).ToString(CultureInfo.InvariantCulture);
-                
+
                 switch (provider.ToLower())
                 {
                     case "memorycacheprovider":
@@ -85,8 +88,12 @@ namespace CacheProvider.Multi
                         break;
                 }
             }
-        }
 
+            _masterCacheProvider = CacheProviders.Last();
+        }
+        #endregion
+
+        #region Get/Exist/Add
         public override async Task<object> Get(object cacheKey, string region)
         {
             if (!_isEnabled)
@@ -94,21 +101,37 @@ namespace CacheProvider.Multi
                 return null;
             }
 
-            foreach (var cp in CacheProviders)
+            // heart beat check to see if this cache item first exist in the final cache.  
+            // this is protect in memory cache from becoming stagnant in a load balaance farm
+            // and all the in memory farm is old
+            if (await _masterCacheProvider.Exist(cacheKey, region))
             {
-                var obj = await cp.Get(cacheKey, region);
-                if (obj != null)
+                foreach (var cp in CacheProviders)
                 {
-                    return obj;
+                    var obj = await cp.Get(cacheKey, region);
+                    if (obj != null)
+                    {
+                        return obj;
+                    }
                 }
             }
 
             return null;
         }
 
+        public override Task<object> Get(object cacheKey, string region, string validationKey)
+        {
+            throw new NotImplementedException();
+        }
+
         public override async Task<T> Get<T>(object cacheKey, string region)
         {
             return (T)await Get(cacheKey, region);
+        }
+
+        public override Task<T> Get<T>(object cacheKey, string region, string validationKey)
+        {
+            throw new NotImplementedException();
         }
 
         public override async Task<bool> Exist(object cacheKey, string region)
@@ -129,6 +152,107 @@ namespace CacheProvider.Multi
             return false;
         }
 
+        public override async Task<long> Count(string region)
+        {
+            if (!_isEnabled)
+            {
+                return 0;
+            }
+
+            long highCount = 0;
+            foreach (var cp in CacheProviders)
+            {
+                var count = await cp.Count(region);
+                if (highCount < count)
+                {
+                    highCount = count;
+                }
+            }
+            return highCount;
+        }
+        #endregion
+
+        #region Add
+
+        /// <summary>
+        ///     Add to cache.
+        /// </summary>
+        /// <param name="cacheKey">The cache key.</param>
+        /// <param name="cacheObject">The cache object.</param>
+        /// <param name="region">If region is supported by cache , it will seperate the lookups</param>
+        /// <param name="options">Options that can be set for the cache</param>
+        /// <returns>True if successful else false.</returns>
+        public override async Task<bool> Add(object cacheKey, object cacheObject, string region, ICacheOptions options)
+        {
+            if (!_isEnabled)
+            {
+                return true;
+            }
+
+            if (string.IsNullOrWhiteSpace(options.Validator))
+            {
+                options.Validator = Guid.NewGuid().ToString();
+            }
+
+            var returnValue = true;
+            foreach (var cp in CacheProviders)
+            {
+                var wasAdded = await cp.Add(cacheKey, cacheObject, region, options);
+                if (!wasAdded)
+                {
+                    returnValue = false;
+                }
+            }
+
+            return returnValue;
+        }
+
+        /// <summary>
+        ///     Add an item to the cache and will need to be removed manually
+        /// </summary>
+        /// <param name="cacheKey">The cache key.</param>
+        /// <param name="cacheObject">The cache object.</param>
+        /// <param name="region">If region is supported by cache , it will seperate the lookups</param>
+        /// <param name="options">Options that can be set for the cache</param>
+        /// <returns>true or false</returns>
+        public override async Task<bool> AddPermanent(object cacheKey, object cacheObject, string region, ICacheOptions options)
+        {
+            if (!_isEnabled)
+            {
+                return true;
+            }
+
+            if (string.IsNullOrWhiteSpace(options.Validator))
+            {
+                options.Validator = Guid.NewGuid().ToString();
+            }
+
+            var returnValue = true;
+            foreach (var cp in CacheProviders)
+            {
+                var wasAdded = await cp.AddPermanent(cacheKey, cacheObject, region, options);
+                if (!wasAdded)
+                {
+                    returnValue = false;
+                }
+            }
+
+            return returnValue;
+        }
+
+        #endregion
+
+        // todo remove after Dec
+        #region Add obsolete
+        /// <summary>
+        ///     Add to cache.
+        /// </summary>
+        /// <param name="cacheKey">The cache key.</param>
+        /// <param name="cacheObject">The cache object.</param>
+        /// <param name="region"></param>
+        /// <param name="expirationInMinutes"></param>
+        /// <returns>True if successful else false.</returns>
+        [Obsolete("will be removed after December, use the other Add with options")]
         public override async Task<bool> Add(object cacheKey, object cacheObject, string region, int expirationInMinutes = 15)
         {
             if (!_isEnabled)
@@ -148,6 +272,16 @@ namespace CacheProvider.Multi
             return returnValue;
         }
 
+        /// <summary>
+        ///     Add to cache.
+        /// </summary>
+        /// <param name="cacheKey">The cache key.</param>
+        /// <param name="cacheObject">The cache object.</param>
+        /// <param name="region"></param>
+        /// <param name="allowSliddingTime">Updates the expiration x minutes from last write or reed</param>
+        /// <param name="expirationInMinutes"></param>
+        /// <returns>True if successful else false.</returns>
+        [Obsolete("will be removed after December, use the other Add with options")]
         public override async Task<bool> Add(object cacheKey, object cacheObject, string region, bool allowSliddingTime, int expirationInMinutes = 15)
         {
             if (!_isEnabled)
@@ -167,6 +301,14 @@ namespace CacheProvider.Multi
             return returnValue;
         }
 
+        /// <summary>
+        ///     Add an item to the cache and will need to be removed manually
+        /// </summary>
+        /// <param name="cacheKey"></param>
+        /// <param name="cacheObject"></param>
+        /// <param name="region"></param>
+        /// <returns>true or false</returns>
+        [Obsolete("will be removed after December, use the other Add with options")]
         public override async Task<bool> AddPermanent(object cacheKey, object cacheObject, string region)
         {
             if (!_isEnabled)
@@ -176,7 +318,7 @@ namespace CacheProvider.Multi
             var returnValue = true;
             foreach (var cp in CacheProviders)
             {
-                var wasAdded = await cp.Add(cacheKey, cacheObject, region);
+                var wasAdded = await cp.AddPermanent(cacheKey, cacheObject, region);
                 if (!wasAdded)
                 {
                     returnValue = false;
@@ -185,7 +327,9 @@ namespace CacheProvider.Multi
 
             return returnValue;
         }
+        #endregion
 
+        #region remove
         public override async Task<bool> Remove(object cacheKey, string region)
         {
             if (!_isEnabled)
@@ -261,24 +405,8 @@ namespace CacheProvider.Multi
 
             return returnValue;
         }
+        #endregion
 
-        public override async Task<long> Count(string region)
-        {
-            if (!_isEnabled)
-            {
-                return 0;
-            }
 
-            long highCount = 0;
-            foreach (var cp in CacheProviders)
-            {
-                var count = await cp.Count(region);
-                if (highCount < count)
-                {
-                    highCount = count;
-                }
-            }
-            return highCount;
-        }
     }
 }
